refactor(website): move interfaces to file
