+++ website/src/pages/downloads/macos.tsx
@@ -5,7 +5,7 @@ import { faBeer, faDownload, faPaste, faTerminal } from '@fortawesome/free-solid
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import TailWindThemeSelector from '@site/src/components/TailWindThemeSelector';
 import { TelemetryLink } from '@site/src/components/TelemetryLink';
-import type { GitHubMetadata } from '@site/src/plugins/github-metadata-plugin'; // Adjust path if needed
+import type { GitHubMetadata } from '@site/src/plugins/GitHubMetadata';
 import Layout from '@theme/Layout';
 import React from 'react';
