+++ website/src/pages/downloads/windows.tsx
@@ -6,7 +6,7 @@ import { faDownload, faPaste, faTerminal } from '@fortawesome/free-solid-svg-ico
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import TailWindThemeSelector from '@site/src/components/TailWindThemeSelector';
 import { TelemetryLink } from '@site/src/components/TelemetryLink';
-import type { GitHubMetadata } from '@site/src/plugins/github-metadata-plugin'; // Adjust path if needed
+import type { GitHubMetadata } from '@site/src/plugins/GitHubMetadata';
 import Layout from '@theme/Layout';
 import React from 'react';
